+++ src/main/kotlin/com/codepied/api/api/externalApi/KakaoLoginApiService.kt
@@ -3,7 +3,8 @@ package com.codepied.api.api.externalApi
 import com.codepied.api.api.config.KakaoSocialLoginProperty
 import com.codepied.api.api.exception.ErrorCode
 import com.codepied.api.api.exception.InvalidRequestExceptionBuilder.invalidRequest
-import com.codepied.api.api.externalApi.dto.SocialAccountImpl
+import com.codepied.api.api.externalApi.dto.KakaoLoginTokenResponse
+import com.codepied.api.api.externalApi.dto.KakaoLoginUserInfo
 import com.codepied.api.api.security.SocialType
 import com.codepied.api.api.security.dto.SocialAccount
 import org.springframework.core.ParameterizedTypeReference
@@ -12,6 +13,12 @@ import org.springframework.http.MediaType
 import org.springframework.stereotype.Service
 import org.springframework.web.reactive.function.client.WebClient
 
+/**
+ * social login Api Service: provider: KAKAO
+ *
+ * @author Aivyss
+ * @since 2022/12/25
+ */
 @Service
 class KakaoLoginApiService(
     private val webClient: WebClient,
@@ -56,38 +63,3 @@ class KakaoLoginApiService(
 
     override fun supportType(): SocialType = SocialType.KAKAO
 }
-
-class KakaoLoginTokenResponse(
-    val token_type: String?,
-    val access_token: String,
-    val id_token: String,
-    val expires_in: Int,
-    val refresh_token: String,
-    val refresh_token_expires_in: Int,
-)
-
-class KakaoLoginUserInfo(
-    val id: Long,
-    val kakao_account: KakaoLoginUserData?
-) {
-    fun create(): SocialAccount {
-        val invalidRequest = invalidRequest(
-            errorCode = ErrorCode.NOT_ACCESSIBLE_SOCIAL_USER_KAKAO,
-            debugMessage = "not valid email"
-        )
-
-        if (kakao_account?.is_email_verified != true) {
-            throw invalidRequest
-        }
-
-        return SocialAccountImpl(
-            id = id.toString(),
-            email = kakao_account?.email ?: throw invalidRequest
-        )
-    }
-}
-
-class KakaoLoginUserData(
-    val is_email_verified: Boolean?,
-    val email: String?
-)
