+++ src/main/kotlin/com/codepied/api/api/externalApi/dto/KakaoLoginUserData.kt
@@ -1,0 +1,6 @@
+package com.codepied.api.api.externalApi.dto
+
+class KakaoLoginUserData(
+    val is_email_verified: Boolean?,
+    val email: String?
+)
